fix spacing and remove dumb "escaped entities" eslint rule
