+++ frontend/new/components/Card/Card.tsx
@@ -1,13 +1,13 @@
 import { CardModel } from '@/lib/models'
 
 interface IProps {
-    card: CardModel,
+  card: CardModel,
 }
 
 export default function Card({ card }: IProps) {
-    return (
-        <div>
-            {card.title}
-        </div>
-    )
+  return (
+    <div>
+      {card.title}
+    </div>
+  )
 }
