+++ frontend/new/components/Header/Header.tsx
@@ -2,11 +2,11 @@ import Link from 'next/link'
 import css from './Header.module.css'
 
 export default function Header() {
-    return (
-        <header className={css.header}>
-            <Link href='/'>
-                <h1 className={css.title}>Kanbad</h1>
-            </Link>
-        </header>
-    )
+  return (
+    <header className={css.header}>
+      <Link href='/'>
+        <h1 className={css.title}>Kanbad</h1>
+      </Link>
+    </header>
+  )
 }
