+++ frontend/new/lib/models.ts
@@ -3,18 +3,18 @@
  */
 
 interface Model {
-    identifier: string,
+  identifier: string,
 }
 
 export interface BoardModel extends Model {
-    workspace: string,
-    title: string,
+  workspace: string,
+  title: string,
 }
 
 export interface CardModel extends Model {
-    board: string,
-    body: string,
-    title: string,
+  board: string,
+  body: string,
+  title: string,
 }
 
 export interface WorkspaceModel extends Model {}
