+++ frontend/new/pages/_app.tsx
@@ -1,5 +1,6 @@
 import '@/styles/globals.css'
 import type { AppProps } from 'next/app'
+import { Head } from 'next/document'
 
 import Header from '@/components/Header/Header'
 
@@ -7,6 +8,9 @@ export default function App({ Component, pageProps }: AppProps) {
 
   return (
     <>
+      <Head>
+        <title>Kanbad</title>
+      </Head>
       <Header />
       <Component {...pageProps} />
     </>
