+++ frontend/new/styles/globals.css
@@ -1,9 +1,9 @@
 :root {
-  background: white;
+    background: white;
 }
 
 * {
-  box-sizing: border-box;
-  padding: 0;
-  margin: 0;
+    box-sizing: border-box;
+    padding: 0;
+    margin: 0;
 }
